выполняет задание "время разбивать камни", разделяет код на модули
