+++ src/components/load-more.js
@@ -1,0 +1,7 @@
+const createLoadMoreTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createLoadMoreTemplate};
